backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
flask-compress==1.14
psycopg2-binary==2.9.6
tensorflow>=2.16.0
numpy>=1.26.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
python-dotenv==1.0.0
tqdm>=4.66.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0